+++ demo/stores/app.js
@@ -36,9 +36,15 @@ let data = ImmutableHelper.parseJSON({
     as: 'info',
     text: 'Pill'
   },
+  portrait: {
+    size: 'lmed',
+    shape: 'standard',
+    email: 'Email'
+  },
   pod: {
     border: true,
-    padding: "medium"
+    padding: "medium",
+    as: "primary"
   },
   rainbow: {
     data: [{
@@ -56,6 +62,10 @@ let data = ImmutableHelper.parseJSON({
   },
   row: {
     columnData: [{}, {}, {}, {}]
+  },
+  sidebar: {
+    open: false,
+    disableBackground: true
   },
   spinner: {
     as: 'info',
@@ -88,9 +98,6 @@ let data = ImmutableHelper.parseJSON({
   },
   tabs: {
     tabData: [{}, {}]
-  },
-  tile: {
-    content: "Using a borderless Pod can provide padding inside your tile."
   },
   toast: {
     as: "warning",