+++ demo/views/modals/sidebar-demo/sidebar-demo.js
@@ -6,48 +6,63 @@ import Example from './../../../components/example';
 import AsDropdown from './../../../components/as-dropdown';
 import SizeDropdown from './../../../components/size-dropdown';
 
-import Dialog from 'components/dialog';
+import { Sidebar, SidebarHeader } from 'components/sidebar';
 import Row from 'components/row';
 import Button from 'components/button';
 import Textbox from 'components/textbox';
 import Checkbox from 'components/checkbox';
 
-class DialogDemo extends React.Component {
+class SidebarDemo extends React.Component {
 
   /**
    * @method value
    */
   value = (key) => {
-    return this.state.appStore.getIn(['dialog', key]);
+    return this.state.appStore.getIn(['sidebar', key]);
   }
 
   /**
    * @method action
    */
   get action() {
-    return AppActions.appValueUpdated.bind(this, 'dialog');
+    return AppActions.appValueUpdated.bind(this, 'sidebar');
   }
 
   /**
    * @method demo
    */
   get demo() {
+    let button;
+
+    if (this.value('open')) {
+      button = (
+        <Button onClick={ this.action.bind(this, 'open', { target: { value: false } } ) } >
+          Click me to Close Sidebar
+        </Button>
+      );
+    } else {
+      button = (
+        <Button onClick={ this.action.bind(this, 'open', { target: { value: true } } ) } >
+          Click me to Open Sidebar 
+        </Button>
+      );
+    }
+
+    let position = this.value('positionLeft') ? 'left' : 'right';
+
     return (
       <div>
-        <Button onClick={ this.action.bind(this, 'open', { target: { value: true } } ) } >
-          Click me to Open Dialog
-        </Button>
-        <Dialog
+        { button }
+        <Sidebar
           open={ this.value('open') }
-          onCancel={ this.action.bind(this, 'open', { target: { value: false } } ) }
-          title={ this.value('title') }
+          onClose={ this.action.bind(this, 'open', { target: { value: false } } ) }
           disableBackground={ this.value('disableBackground') }
+          position={ position }
         >
-          <Row>
-            <Textbox />
-            <Textbox />
-          </Row>
-        </Dialog>
+          <SidebarHeader>
+            Sean Armstrong
+          </SidebarHeader>
+        </Sidebar>
       </div>
     );
   }
@@ -56,16 +71,18 @@ class DialogDemo extends React.Component {
    * @method code
    */
   get code() {
-    let html = "import Dialog from 'carbon/lib/components/dialog';\n\n";
+    let html = "import Sidebar from 'carbon/lib/components/sidebar';\n\n";
 
-    html += "<Dialog\n";
+    html += "<Sidebar\n";
     html += `  open={ ${ this.value('open') } }\n`
-    html += `  title="${ this.value('title') }"\n`;
 
     if (!this.value('disableBackground')) {
-      html += `  disableBackground={ false }\n`;
+      html += `  disableBackground={ false }\n`
     }
 
+    if (this.value('positionLeft')) {
+      html += `  position='left'\n`
+    }
     html += "/>\n\n";
 
     return html;
@@ -78,17 +95,17 @@ class DialogDemo extends React.Component {
     return (
       <div>
         <Row>
-          <Textbox
-            label="Title"
-            labelInline={ true }
-            value={ this.value('title') }
-            onChange={ this.action.bind(this, 'title') }
-          />
           <Checkbox
             label="Disable Background"
             value={ this.value('disableBackground') }
             reverse={ true }
             onChange={ this.action.bind(this, 'disableBackground') }
+          />
+          <Checkbox
+            label="Position Left"
+            value={ this.value('positionLeft') }
+            reverse={ true }
+            onChange={ this.action.bind(this, 'positionLeft') }
           />
         </Row>
       </div>
@@ -101,8 +118,8 @@ class DialogDemo extends React.Component {
   render() {
     return (
       <Example
-        title="Dialog"
-        readme="components/dialog"
+        title="Sidebar"
+        readme="components/sidebar"
         demo={ this.demo }
         code={ this.code }
         controls={ this.controls }
@@ -111,4 +128,4 @@ class DialogDemo extends React.Component {
   }
 }
 
-export default connect(DialogDemo, AppStore);
+export default connect(SidebarDemo, AppStore);