+++ src/components/pod/__spec__.js
@@ -94,14 +94,20 @@ describe('Pod', () => {
     it('renders a parent div with a pod CSS class', () => {
       instance = TestUtils.renderIntoDocument(<Pod/>);
       let podNode = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'div')[0]
-      expect(podNode.className).toEqual('ui-pod ui-pod--padding-medium');
+      expect(podNode.className).toEqual('ui-pod ui-pod--primary ui-pod--padding-medium');
+    });
+
+    it('renders a parent div with an as CSS class', () => {
+      instance = TestUtils.renderIntoDocument(<Pod as="secondary"/>);
+      let podNode = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'div')[0]
+      expect(podNode.className).toEqual('ui-pod ui-pod--secondary ui-pod--padding-medium');
     });
 
     describe('when a custom className is passed', () => {
       it('adds the class to the surrounding parent div', () => {
         instance = TestUtils.renderIntoDocument(<Pod className="CustomClass"/>);
         let podNode = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'div')[0]
-        expect(podNode.className).toEqual('ui-pod CustomClass ui-pod--padding-medium');
+        expect(podNode.className).toEqual('ui-pod CustomClass ui-pod--primary ui-pod--padding-medium');
       });
     });
 
@@ -109,7 +115,7 @@ describe('Pod', () => {
       it('adds class to disable border', () => {
         instance = TestUtils.renderIntoDocument(<Pod border={ false }/>);
         let podNode = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'div')[0]
-        expect(podNode.className).toEqual('ui-pod ui-pod--padding-medium ui-pod--no-border');
+        expect(podNode.className).toEqual('ui-pod ui-pod--primary ui-pod--padding-medium ui-pod--no-border');
       });
     });
 
@@ -117,7 +123,7 @@ describe('Pod', () => {
       it('adds class to use padding size', () => {
         instance = TestUtils.renderIntoDocument(<Pod padding="small" />);
         let podNode = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'div')[0]
-        expect(podNode.className).toEqual('ui-pod ui-pod--padding-small');
+        expect(podNode.className).toEqual('ui-pod ui-pod--primary ui-pod--padding-small');
       });
     });